requirements_day10.txt
# Day 10 Dashboard Dependencies
# EyeD AI Attendance System

# Core dashboard framework
streamlit>=1.28.0

# Data processing and analysis
pandas>=1.5.0
numpy>=1.21.0

# Visualization and charts
plotly>=5.15.0

# Image processing
opencv-python>=4.8.0
Pillow>=9.0.0

# Date and time handling
python-dateutil>=2.8.0

# Optional: Enhanced features
# streamlit-aggrid>=0.3.0  # For advanced tables
# streamlit-option-menu>=0.3.0  # For better navigation
# streamlit-authenticator>=0.2.0  # For authentication (future use)

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
